+++ src/integrations/elements/User.php
@@ -211,7 +211,11 @@ class User extends Element
             }
 
             $attributeValues = $this->getFieldMappingValues($submission, $this->attributeMapping, $this->getElementAttributes());
-            $attributeValues = array_filter($attributeValues);
+            
+            // Filter null values
+            $attributeValues = array_filter($attributeValues, function($var) {
+                return $var !== null;
+            });
 
             // Check if the password was mapped, as if the source field was a Password field.
             // The value will already be hashed and we need to do a manual DB-level update
@@ -227,7 +231,11 @@ class User extends Element
 
             $fields = $this->getFormSettingValue('elements')[0]->fields ?? [];
             $fieldValues = $this->getFieldMappingValues($submission, $this->fieldMapping, $fields);
-            $fieldValues = array_filter($fieldValues);
+            
+            // Filter null values
+            $fieldValues = array_filter($fieldValues, function($var) {
+                return $var !== null;
+            });
 
             $user->setFieldValues($fieldValues);